sys/log: add format attribute to modlog_printf prototype (#3530)
- Add format(printf) attribute to modlog_printf to enable compile-time
checking that format strings match arguments. By default the
attribute is disabled; it can be enabled by setting
MODLOG_USE_PRINTF_ATTRIBUTE to 1.
Note: Enabling the attribute will cause incorrect modlog_printf
usage which previously compiled successfully, to no longer compile.

+++ net/oic/src/port/mynewt/ip_adaptor.c
@@ -165,8 +165,9 @@ oc_send_buffer_ip6_int(struct os_mbuf *m, int is_mcast)
             to.msin6_scope_id = itf2.mif_idx;
             rc = mn_sendto(oc_ucast6, n, (struct mn_sockaddr *) &to);
             if (rc != 0) {
-                OC_LOG_ERROR("Failed to send buffer %u on itf %d\n",
-                             OS_MBUF_PKTHDR(m)->omp_len, to.msin6_scope_id);
+                OC_LOG_ERROR("Failed to send buffer %u on itf %u\n",
+                             OS_MBUF_PKTHDR(m)->omp_len,
+                             (unsigned int)to.msin6_scope_id);
                 STATS_INC(oc_ip_stats, oerr);
                 os_mbuf_free_chain(n);
             }
@@ -176,8 +177,9 @@ oc_send_buffer_ip6_int(struct os_mbuf *m, int is_mcast)
             to.msin6_scope_id = itf2.mif_idx;
             rc = mn_sendto(oc_ucast6, m, (struct mn_sockaddr *) &to);
             if (rc != 0) {
-                OC_LOG_ERROR("Failed sending buffer %u on itf %d\n",
-                             OS_MBUF_PKTHDR(m)->omp_len, to.msin6_scope_id);
+                OC_LOG_ERROR("Failed sending buffer %u on itf %u\n",
+                             OS_MBUF_PKTHDR(m)->omp_len,
+                             (unsigned int)to.msin6_scope_id);
                 STATS_INC(oc_ip_stats, oerr);
                 os_mbuf_free_chain(m);
             }
@@ -187,8 +189,8 @@ oc_send_buffer_ip6_int(struct os_mbuf *m, int is_mcast)
     } else {
         rc = mn_sendto(oc_ucast6, m, (struct mn_sockaddr *) &to);
         if (rc != 0) {
-            OC_LOG_ERROR("Failed to send buffer %u on itf %d\n",
-                         OS_MBUF_PKTHDR(m)->omp_len, to.msin6_scope_id);
+            OC_LOG_ERROR("Failed to send buffer %u on itf %u\n",
+                         OS_MBUF_PKTHDR(m)->omp_len, (unsigned int)to.msin6_scope_id);
             STATS_INC(oc_ip_stats, oerr);
             os_mbuf_free_chain(m);
         }
